赛道管理-增删改查

+++ modules/transfer-crm/src/main/java/com/hqjy/mustang/transfer/crm/model/dto/TransferCustomerDTO.java
@@ -64,7 +64,7 @@ public class TransferCustomerDTO {
     /**
      * 性别:见数据字典SEX枚举(0-未知，1-男，2-女) sex
      **/
-    private Byte sex;
+    private Integer sex;
 
     /**
      * 年龄 age
